telegram-bot/requirements.txt
aio-pika==9.4.3
httpx==0.27.2
python-dotenv==1.0.1
